--- OneDay/OneDay/Controller/Entry/CreateEntryPhotosViewController.swift
@@ -1,30 +1,0 @@
-//
-//  CreateEntryPhotosViewController.swift
-//  OneDay
-//
-//  Created by juhee on 11/02/2019.
-//  Copyright © 2019 teamA2. All rights reserved.
-//
-
-import UIKit
-
-class CreateEntryPhotosViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
