+++ OneDay/OneDay/Controller/MainTabBarViewController.swift
@@ -53,61 +53,3 @@ extension MainTabBarViewController: UITabBarControllerDelegate, UIImagePickerCon
     }
 
 }
-
-class AddActionViewController: UIViewController {
-    
-}
-extension AddActionViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-    
-    func imagePickerController(_ picker: UIImagePickerController,
-                               didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]
-        ) {
-        picker.dismiss(animated: true, completion: nil)
-        createEntryWithImage(pickingMediaWithInfo: info)
-    }
-}
-
-enum ImageSource {
-    case photoLibrary
-    case camera
-}
-
-extension UIImagePickerControllerDelegate where Self: UIViewController, Self: UINavigationControllerDelegate {
-    
-    func selectImageFrom(_ source: ImageSource) {
-        let imagePicker =  UIImagePickerController()
-        imagePicker.delegate = self
-        switch source {
-        case .camera:
-            imagePicker.sourceType = .camera
-        case .photoLibrary:
-            imagePicker.sourceType = .photoLibrary
-            imagePicker.allowsEditing = true
-        }
-        present(imagePicker, animated: true, completion: nil)
-    }
-    
-    func createEntryWithImage(pickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        
-        var pickedImage: UIImage!
-        if let editedImage = info[.editedImage] as? UIImage {
-            pickedImage = editedImage
-        } else if let originImage = info[.originalImage] as? UIImage {
-            pickedImage = originImage
-        } else {
-            print("Image not found!")
-            return
-        }
-        
-        let originWidth = pickedImage.size.width
-        let scaleFactor = originWidth / (view.frame.size.width - 10)
-        let scaledImage =  UIImage(cgImage: pickedImage.cgImage!, scale: scaleFactor, orientation: .up)
-        
-        guard let nextVC = UIStoryboard(name: "Coredata", bundle: nil).instantiateViewController(withIdentifier: "entry_detail") as? EntryViewController else { return }
-        let newEntry: Entry = CoreDataManager.shared.insert()
-        newEntry.contents = scaledImage.attributedString
-        nextVC.entry = newEntry
-        present(nextVC, animated: true)
-    }
-    
-}
