+++ OneDay/OneDay/Model/ImageSource.swift
@@ -1,0 +1,14 @@
+//
+//  ImageSource.swift
+//  OneDay
+//
+//  Created by juhee on 14/02/2019.
+//  Copyright © 2019 teamA2. All rights reserved.
+//
+
+import Foundation
+
+enum ImageSource {
+    case photoLibrary
+    case camera
+}
